requirements.txt
langchain
langchain_community
pytube
langchain_openai
python-dotenv
youtube_transcript_api
google-generativeai
